@@ -1,101 +1,74 @@
 {
     "name": "Korp",
-    "version": "9.1.0",
+    "version": "9.3.0",
     "dependencies": {
-        "angular": "~1.7.0",
-        "angular-dynamic-locale": "~0.1.37",
+        "@fortawesome/fontawesome-free": "6.2.1",
+        "angular": "1.8.3",
+        "angular-dynamic-locale": "0.1.38",
         "angular-filter": "0.5.17",
         "angular-spinner": "1.0.1",
         "angular-ui-bootstrap": "2.5.6",
         "angular-ui-sortable": "~0.19.0",
-        "bootstrap-sass": "^3.3.7",
+        "bootstrap-sass": "^3.4.3",
         "comma-separated-values": "3.6.4",
         "components-jqueryui": "1.12.1",
-        "font-awesome": "4.3.0",
         "geokorp": "spraakbanken/geokorp#1.4.2",
-        "jquery": "3.2.1",
+        "jquery": "3.6.3",
         "jquery-deparam": "^0.5.3",
         "jquery-flot": "0.8.3",
-        "json3": "3.3.2",
+        "json3": "3.3.3",
         "jstorage": "andris9/jStorage#0.4.12",
-        "leaflet": "^1.6.0",
-        "leaflet-providers": "^1.9.1",
-        "leaflet.markercluster": "^1.4.1",
-        "lodash": "^4.17.19",
-        "moment": "2.20.1",
-        "raphael": "2.2.7",
-        "raw-loader": "^4.0.2",
-        "rickshaw": "1.6.4",
-        "slickgrid": "spraakbanken/SlickGrid#master",
-        "tailwindcss": "^2.0.3",
-        "xstate": "^4.13.0"
+        "leaflet": "^1.9.3",
+        "leaflet-providers": "^1.13.0",
+        "leaflet.markercluster": "^1.5.3",
+        "lodash": "^4.17.21",
+        "moment": "2.29.4",
+        "raphael": "2.3.0",
+        "rickshaw": "1.7.1",
+        "slickgrid": "3.0.3",
+        "tailwindcss": "3.2.4",
+        "xstate": "4.35.4"
     },
     "devDependencies": {
-        "@babel/core": "^7.2.0",
-        "@babel/plugin-transform-runtime": "^7.2.0",
-        "@babel/preset-env": "^7.2.0",
-        "@babel/runtime": "^7.2.0",
         "@types/lodash": "^4.14.118",
         "autoprefixer": "^10.2.4",
-        "babel-loader": "^8.0.5",
-        "babel-plugin-angularjs-annotate": "^0.9.0",
-        "babel-plugin-syntax-dynamic-import": "^6.18.0",
-        "clean-webpack-plugin": "^3.0.0",
-        "compression-webpack-plugin": "^4.0.0",
-        "copy-webpack-plugin": "^6.0.3",
-        "css-hot-loader": "^1.3.9",
-        "css-loader": "^5.1.1",
-        "eslint": "^7.8.1",
-        "eslint-config-standard": "^14.1.1",
-        "eslint-plugin-import": "^2.22.0",
-        "eslint-plugin-node": "^11.1.0",
-        "eslint-plugin-promise": "^4.2.1",
-        "eslint-plugin-standard": "^4.0.1",
-        "extract-loader": "^1.0.2",
-        "file-loader": "^1.1.11",
-        "html-loader": "^0.5.5",
-        "http-server": "^0.11.1",
-        "imports-loader": "^0.8.0",
-        "jasmine-core": "^2.99.1",
-        "karma": "^5.1.1",
-        "karma-chrome-launcher": "^3.1.0",
-        "karma-jasmine": "^3.3.1",
-        "karma-webpack": "^4.0.2",
-        "pegjs": "^0.10.0",
-        "postcss-loader": "^4.2.0",
+        "clean-webpack-plugin": "^4.0.0",
+        "compression-webpack-plugin": "^10.0.0",
+        "copy-webpack-plugin": "^11.0.0",
+        "css-loader": "^6.0.0",
+        "html-loader": "^4.0.0",
+        "jasmine-core": "^4.0.0",
+        "postcss-loader": "^7.0.0",
         "prettier": "^2.0.5",
         "protractor": "^5.4.2",
-        "pug-html-loader": "^1.1.5",
-        "pug-loader": "^2.4.0",
-        "sass": "~1.32.12",
-        "sass-loader": "^10.1.1",
-        "style-loader": "^2.0.0",
-        "ts-loader": "^5.3.1",
-        "tslint": "^6.1.3",
-        "tslint-config-prettier": "^1.17.0",
-        "typescript": "^3.3.3333",
-        "wait-on": "^5.2.0",
-        "webpack": "~4.44.1",
-        "webpack-bundle-analyzer": "^3.8.0",
-        "webpack-cli": "^3.3.12",
-        "webpack-dev-server": "^3.11.0",
-        "webpack-merge": "^5.1.1"
+        "sass": "~1.49.9",
+        "sass-loader": "^13.0.0",
+        "style-loader": "^3.0.0",
+        "ts-loader": "^9.0.0",
+        "typescript": "^4.0.0",
+        "wait-on": "^7.0.0",
+        "webpack": "^5.0.0",
+        "webpack-bundle-analyzer": "^4.0.0",
+        "webpack-cli": "^5.0.0",
+        "webpack-dev-server": "^4.0.0",
+        "webpack-merge": "^5.1.1",
+        "yaml-loader": "^0.8.0"
     },
     "engines": {
-        "node": ">=0.10.0"
+        "node": ">=14"
     },
     "scripts": {
-        "test": "yarn run test:karma && yarn run test:e2e",
-        "actions-test": "yarn run test:karma && yarn run test:e2e",
+        "test": "yarn run test:e2e",
+        "actions-test": "yarn run test:e2e",
         "posttest": "kill $(lsof -t -i:9112)",
-        "test:karma": "karma start test/karma/karma.conf.js",
         "pretest:e2e": "yarn run testserver && node_modules/protractor/bin/webdriver-manager update --standalone chrome --gecko false && wait-on --delay 500 --interval 500 http://localhost:9112",
-        "test:e2e": "node_modules/protractor/bin/protractor --browser chrome test/e2e/conf_e2e.js || true",
+        "test:e2e": "KORP_HOST=localhost node_modules/protractor/bin/protractor --browser chrome test/e2e/conf_e2e.js || true",
         "start": "webpack-dev-server --config webpack.dev.js",
+        "start:labb": "NODE_ENV=staging webpack-dev-server --config webpack.dev.js",
         "dev": "webpack-dev-server --config webpack.dev.js",
-        "testserver": "webpack-dev-server --config webpack.dev.js --port 9112 &",
-        "start:dist": "http-server -p ${KORP_PORT:-9111} dist/",
+        "testserver": "KORP_HOST=localhost KORP_HTTPS= webpack-dev-server --config webpack.dev.js --port 9112 &",
         "build": "NODE_ENV=production webpack --config webpack.prod.js",
+        "build:labb": "NODE_ENV=staging webpack --config webpack.prod.js",
         "lint": "tslint --project tslint.json"
     }
 }
