@@ -1,14 +1,41 @@
 /** @format */
-
-import jStorage from "../lib/jstorage"
 import { kwicPagerName, kwicPager } from "./components/pager"
 import { sidebarName, sidebarComponent } from "./components/sidebar"
 import * as autoc from "./components/autoc"
 import * as readingmode from "./components/readingmode"
 import * as extendedAddBox from "./components/extended/extended_add_box"
-import { setDefaultConfigValues } from "./settings.js"
+import { corpusChooserComponent } from "./components/corpus_chooser/corpus_chooser"
+import { ccTimeGraphComponent } from "./components/corpus_chooser/time_graph"
+import { ccTreeComponent } from "./components/corpus_chooser/tree"
+import { ccInfoBox } from "./components/corpus_chooser/info_box"
+import { depTreeComponent } from "./components/deptree/deptree"
+import { simpleSearchComponent } from "./components/simple_search"
+import { extendedStandardComponent } from "./components/extended/standard_extended"
+import { extendedParallelComponent } from "./components/extended/parallel_extended"
+import { extendedTokensComponent } from "./components/extended/extended_tokens"
+import { extendedAndTokenComponent } from "./components/extended/and_token"
+import { extendedCQPTermComponent } from "./components/extended/cqp_term"
+import { extendedTokenComponent } from "./components/extended/token"
+import { extendedStructTokenComponent } from "./components/extended/struct_token"
+import { extendedCQPValueComponent } from "./components/extended/cqp_value"
+import { advancedSearchComponent } from "./components/advanced_search"
+import { compareSearchComponent } from "./components/compare_search"
+import { kwicComponent } from "./components/kwic"
+import { statisticsComponent } from "./components/statistics"
+import { trendDiagramComponent } from "./components/trend_diagram"
+import { korpErrorComponent } from "./components/korp_error"
+import { kwicTabsComponent } from "./components/dynamic_tabs/kwic_tabs"
+import { graphTabsComponent } from "./components/dynamic_tabs/graph_tabs"
+import { compareTabsComponent } from "./components/dynamic_tabs/compare_tabs"
+import { mapTabsComponent } from "./components/dynamic_tabs/map_tabs"
+import { textTabsComponent } from "./components/dynamic_tabs/text_tabs"
+import { headerComponent } from "./components/header"
+import { wordPictureComponent } from "./components/word_picture"
+import { searchtabsComponent } from "./components/searchtabs"
+import { resultsComponent } from "./components/results"
+import statemachine from "@/statemachine"
 
-setDefaultConfigValues()
+let html = String.raw
 
 window.korpApp = angular.module("korpApp", [
     "ui.bootstrap.typeahead",
@@ -44,10 +71,41 @@ window.korpApp = angular.module("korpApp", [
     "angular.filter",
 ])
 
-korpApp.component(kwicPagerName, kwicPager).component(sidebarName, sidebarComponent)
+korpApp.component(kwicPagerName, kwicPager)
+korpApp.component(sidebarName, sidebarComponent)
 korpApp.component(readingmode.componentName, readingmode.component)
 korpApp.component(autoc.componentName, autoc.component)
 korpApp.component(extendedAddBox.componentName, extendedAddBox.component)
+korpApp.component("corpusChooser", corpusChooserComponent)
+korpApp.component("ccTimeGraph", ccTimeGraphComponent)
+korpApp.component("ccTree", ccTreeComponent)
+korpApp.component("ccInfoBox", ccInfoBox)
+korpApp.component("depTree", depTreeComponent)
+korpApp.component("simpleSearch", simpleSearchComponent)
+korpApp.component("extendedStandard", extendedStandardComponent)
+korpApp.component("extendedParallel", extendedParallelComponent)
+korpApp.component("extendedTokens", extendedTokensComponent)
+korpApp.component("extendedAndToken", extendedAndTokenComponent)
+korpApp.component("extendedCqpTerm", extendedCQPTermComponent)
+korpApp.component("extendedToken", extendedTokenComponent)
+korpApp.component("extendedStructToken", extendedStructTokenComponent)
+korpApp.component("extendedCqpValue", extendedCQPValueComponent)
+korpApp.component("advancedSearch", advancedSearchComponent)
+korpApp.component("compareSearch", compareSearchComponent)
+korpApp.component("kwic", kwicComponent)
+korpApp.component("statistics", statisticsComponent)
+korpApp.component("trendDiagram", trendDiagramComponent)
+korpApp.component("korpError", korpErrorComponent)
+korpApp.component("header", headerComponent)
+korpApp.component("wordPicture", wordPictureComponent)
+korpApp.component("searchtabs", searchtabsComponent)
+korpApp.component("results", resultsComponent)
+// these are directives because it needs replace: true, which is not supported in component
+korpApp.directive("kwicTabs", kwicTabsComponent)
+korpApp.directive("graphTabs", graphTabsComponent)
+korpApp.directive("compareTabs", compareTabsComponent)
+korpApp.directive("mapTabs", mapTabsComponent)
+korpApp.directive("textTabs", textTabsComponent)
 
 // load all custom components
 let customComponents = {}
@@ -61,338 +119,242 @@ for (const componentName in customComponents) {
     korpApp.component(componentName, customComponents[componentName])
 }
 
-korpApp.config((tmhDynamicLocaleProvider) =>
-    tmhDynamicLocaleProvider.localeLocationPattern("translations/angular-locale_{{locale}}.js")
-)
+korpApp.filter("mapper", () => (item, f) => f(item))
+korpApp.filter("loc", () => util.getLocaleString)
+korpApp.filter("locObj", () => util.getLocaleStringObject)
+korpApp.filter("replaceEmpty", function () {
+    return function (input) {
+        if (input === "") {
+            return "–"
+        } else {
+            return input
+        }
+    }
+})
 
-korpApp.config(($uibTooltipProvider) =>
-    $uibTooltipProvider.options({
-        appendToBody: true,
-    })
-)
+authenticationProxy.initAngular()
 
-korpApp.config(["$locationProvider", ($locationProvider) => $locationProvider.hashPrefix("")])
-
+/**
+ * angular-dynamic-locale updates translations in the builtin $locale service, which is used
+ * by at least the datepicker in angular-ui-bootstrap.
+ */
 korpApp.config([
-    "$compileProvider",
-    ($compileProvider) => $compileProvider.aHrefSanitizationWhitelist(/^\s*(https?|ftp|mailto|tel|file|blob):/),
+    "tmhDynamicLocaleProvider",
+    (tmhDynamicLocaleProvider) =>
+        tmhDynamicLocaleProvider.localeLocationPattern("translations/angular-locale_{{locale}}.js"),
 ])
 
-korpApp.run(function ($rootScope, $location, searches, tmhDynamicLocale, $q, $timeout) {
-    const s = $rootScope
-    s._settings = settings
-    window.lang = s.lang = $location.search().lang || settings.defaultLanguage
-    s.word_selected = null
-    s.isLab = window.isLab
+korpApp.config([
+    "$uibTooltipProvider",
+    ($uibTooltipProvider) =>
+        $uibTooltipProvider.options({
+            appendToBody: true,
+        }),
+])
 
-    // s.sidebar_visible = false
+/**
+ * Makes the hashPrefix "" instead of "!" which means our URL:s are ?mode=test#?lang=eng
+ * instead of ?mode=test#!?lang=eng
+ */
+korpApp.config(["$locationProvider", ($locationProvider) => $locationProvider.hashPrefix("")])
 
-    s.extendedCQP = null
+/**
+ * "blob" must be added to the trusted URL:s, otherwise downloading KWIC and statistics etc. will not work
+ */
+korpApp.config([
+    "$compileProvider",
+    ($compileProvider) => $compileProvider.aHrefSanitizationTrustedUrlList(/^\s*(https?|ftp|mailto|tel|file|blob):/),
+])
 
-    s.globalFilterDef = $q.defer()
+korpApp.run([
+    "$rootScope",
+    "$location",
+    "tmhDynamicLocale",
+    "$q",
+    "$timeout",
+    "$uibModal",
+    function ($rootScope, $location, tmhDynamicLocale, $q, $timeout, $uibModal) {
+        const s = $rootScope
+        s._settings = settings
+        window.lang = s.lang = $location.search().lang || settings["default_language"]
 
-    s.locationSearch = function () {
-        const search = $location.search(...arguments)
-        $location.replace()
-        return search
-    }
+        s.isLab = window.isLab
 
-    s.searchtabs = () => $(".search_tabs > ul").scope().tabset.tabs
+        s.extendedCQP = null
 
-    tmhDynamicLocale.set("en")
+        s.globalFilterDef = $q.defer()
 
-    s._loc = $location
+        s.locationSearch = function () {
+            const search = $location.search(...arguments)
+            $location.replace()
+            return search
+        }
 
-    s.$watch("_loc.search()", function () {
-        _.defer(() => (window.onHashChange || _.noop)())
+        s.searchtabs = () => $(".search_tabs > ul").scope().tabset.tabs
 
-        return tmhDynamicLocale.set($location.search().lang || "sv")
-    })
+        s._loc = $location
 
-    $rootScope.kwicTabs = []
-    $rootScope.compareTabs = []
-    $rootScope.graphTabs = []
-    $rootScope.mapTabs = []
-    $rootScope.textTabs = []
+        s.$watch("_loc.search()", function () {
+            _.defer(() => (window.onHashChange || _.noop)())
+            tmhDynamicLocale.set($location.search().lang || settings["default_language"])
+        })
 
-    if ($location.search().corpus) {
-        const initialCorpora = []
+        $(document).keyup(function (event) {
+            if (event.keyCode === 27) {
+                $rootScope.$broadcast("abort_requests")
+            }
+        })
 
-        function findInFolder(folder) {
-            // checks if folder is an actual folder of corpora and recursively
-            // collects all corpora in this folder and subfolders
-            const corpusIds = []
-            if (folder && folder.contents) {
-                for (let corpusId of folder.contents) {
-                    corpusIds.push(corpusId)
-                }
-                for (let subFolderId of Object.keys(folder)) {
-                    for (let corpusId of findInFolder(folder[subFolderId])) {
-                        corpusIds.push(corpusId)
+        $rootScope.kwicTabs = []
+        $rootScope.compareTabs = []
+        $rootScope.graphTabs = []
+        $rootScope.mapTabs = []
+        $rootScope.textTabs = []
+
+        s.waitForLogin = false
+
+        function initialzeCorpusSelection(selectedIds) {
+            let loginNeededFor = []
+
+            for (let corpusId of selectedIds) {
+                const corpusObj = settings.corpora[corpusId]
+                if (corpusObj && corpusObj["limited_access"]) {
+                    if (!authenticationProxy.hasCredential(corpusId.toUpperCase())) {
+                        loginNeededFor.push(corpusObj)
                     }
                 }
             }
-            return corpusIds
-        }
 
-        for (let corpus of $location.search().corpus.split(",")) {
-            const corpusObj = settings.corpusListing.struct[corpus]
-            if (corpusObj) {
-                initialCorpora.push(corpusObj)
-            } else {
-                // corpus does not correspond to a corpus ID, check if it is a folder
-                for (let folderCorpus of findInFolder(settings.corporafolders[corpus])) {
-                    if (settings.corpusListing.struct[folderCorpus]) {
-                        initialCorpora.push(settings.corpusListing.struct[folderCorpus])
+            const allCorpusIds = settings.corpusListing.corpora.map((corpus) => corpus.id)
+
+            if (settings["initalization_checks"] && settings["initalization_checks"](s)) {
+                // something was triggered
+            } else if (_.isEmpty(settings.corpora)) {
+                s.openErrorModal({
+                    content: "<korp-error></korp-error>",
+                    resolvable: false,
+                })
+            } else if (loginNeededFor.length != 0) {
+                const loginNeededHTML = () =>
+                    loginNeededFor
+                        .map((corpus) => `<span>${util.getLocaleStringObject(corpus.title)}</span>`)
+                        .join(", ")
+
+                if (authenticationProxy.isLoggedIn()) {
+                    if (settings.corpusListing.corpora.length == loginNeededFor.length) {
+                        s.openErrorModal({
+                            content: "{{'access_denied' | loc:lang}}",
+                            buttonText: "go_to_start",
+                            onClose: () => {
+                                window.location.href = window.location.href.split("?")[0]
+                            },
+                        })
+                    } else {
+                        s.openErrorModal({
+                            content: html`<div>{{'access_partly_denied' | loc:lang}}:</div>
+                                <div>${loginNeededHTML()}</div>
+                                <div>{{'access_partly_denied_continue' | loc:lang}}</div>`,
+                            onClose: () => {
+                                const neededIds = loginNeededFor.map((corpus) => corpus.id)
+                                let newIds = selectedIds.filter((corpusId) => !neededIds.includes(corpusId))
+                                if (newIds.length == 0) {
+                                    newIds = settings["preselected_corpora"]
+                                }
+                                initialzeCorpusSelection(newIds)
+                            },
+                        })
                     }
-                }
-            }
-        }
-
-        const loginNeededFor = []
-        for (let corpusObj of initialCorpora) {
-            if (corpusObj.limitedAccess) {
-                if (
-                    _.isEmpty(authenticationProxy.loginObj) ||
-                    !authenticationProxy.loginObj.credentials.includes(corpusObj.id.toUpperCase())
-                ) {
-                    loginNeededFor.push(corpusObj)
-                }
-            }
-        }
-        s.loginNeededFor = loginNeededFor
-
-        if (!_.isEmpty(s.loginNeededFor)) {
-            s.savedState = $location.search()
-            $location.url($location.path() + "?")
-
-            // some state need special treatment
-            if (s.savedState.reading_mode) {
-                $location.search("reading_mode")
-            }
-            if (s.savedState.search_tab) {
-                $location.search("search_tab", s.savedState.search_tab)
-            }
-            if (s.savedState.cqp) {
-                $location.search("cqp", s.savedState.cqp)
-            }
-
-            $location.search("display", "login")
-        }
-    }
-
-    s.restorePreLoginState = function () {
-        if (s.savedState) {
-            for (let key in s.savedState) {
-                const val = s.savedState[key]
-                if (key !== "search_tab") {
-                    $location.search(key, val)
-                }
-            }
-
-            // some state need special treatment
-            s.$broadcast("updateAdvancedCQP")
-
-            const corpora = s.savedState.corpus.split(",")
-            settings.corpusListing.select(corpora)
-            corpusChooserInstance.corpusChooser("selectItems", corpora)
-
-            s.savedState = null
-            s.loginNeededFor = null
-        }
-    }
-
-    s.searchDisabled = false
-    s.$on("corpuschooserchange", function (event, corpora) {
-        settings.corpusListing.select(corpora)
-        $location.search("corpus", corpora.join(","))
-        s.searchDisabled = settings.corpusListing.selected.length === 0
-    })
-
-    searches.infoDef.then(function () {
-        let { corpus } = $location.search()
-        let currentCorpora = []
-        if (corpus) {
-            currentCorpora = _.flatten(
-                _.map(corpus.split(","), (val) => getAllCorporaInFolders(settings.corporafolders, val))
-            )
-        } else {
-            if (!(settings.preselectedCorpora && settings.preselectedCorpora.length)) {
-                currentCorpora = _.map(settings.corpusListing.corpora, "id")
-            } else {
-                for (let pre_item of settings.preselectedCorpora) {
-                    pre_item = pre_item.replace(/^__/g, "")
-                    currentCorpora = [].concat(
-                        currentCorpora,
-                        getAllCorporaInFolders(settings.corporafolders, pre_item)
-                    )
-                }
-            }
-
-            settings.preselectedCorpora = currentCorpora
-        }
-
-        settings.corpusListing.select(currentCorpora)
-        corpusChooserInstance.corpusChooser("selectItems", currentCorpora)
-    })
-})
-
-korpApp.controller("headerCtrl", function ($scope, $uibModal, utils) {
-    const s = $scope
-
-    s.logoClick = function () {
-        const [baseUrl, modeParam, langParam] = $scope.getUrlParts(currentMode)
-        window.location = baseUrl + modeParam + langParam
-        if (langParam.length > 0) {
-            window.location.reload()
-        }
-    }
-
-    s.citeClick = () => {
-        s.show_modal = "about"
-    }
-
-    s.showLogin = () => {
-        s.show_modal = "login"
-    }
-
-    s.logout = function () {
-        authenticationProxy.loginObj = {}
-        jStorage.deleteKey("creds")
-
-        // TODO figure out another way to do this
-        for (let corpusObj of settings.corpusListing.corpora) {
-            const corpus = corpusObj.id
-            if (corpusObj.limitedAccess) {
-                $(`#hpcorpus_${corpus}`).closest(".boxdiv").addClass("disabled")
-            }
-        }
-        $("#corpusbox").corpusChooser("updateAllStates")
-
-        let newCorpora = []
-        for (let corpus of settings.corpusListing.getSelectedCorpora()) {
-            if (!settings.corpora[corpus].limitedAccess) {
-                newCorpora.push(corpus)
-            }
-        }
-
-        if (_.isEmpty(newCorpora)) {
-            newCorpora = settings.preselectedCorpora
-        }
-        settings.corpusListing.select(newCorpora)
-        s.loggedIn = false
-        $("#corpusbox").corpusChooser("selectItems", newCorpora)
-    }
-
-    const N_VISIBLE = settings.visibleModes
-
-    s.modes = _.filter(settings.modeConfig)
-    if (!isLab) {
-        s.modes = _.filter(settings.modeConfig, (item) => item.labOnly !== true)
-    }
-
-    s.visible = s.modes.slice(0, N_VISIBLE)
-    s.menu = s.modes.slice(N_VISIBLE)
-
-    const i = _.map(s.menu, "mode").indexOf(currentMode)
-    if (i !== -1) {
-        s.visible.push(s.menu[i])
-        s.menu.splice(i, 1)
-    }
-
-    for (let mode of s.modes) {
-        mode.selected = false
-        if (mode.mode === currentMode) {
-            window.settings.mode = mode
-            mode.selected = true
-        }
-    }
-
-    s.getUrl = function (modeId) {
-        return s.getUrlParts(modeId).join("")
-    }
-
-    s.getUrlParts = function (modeId) {
-        const langParam = settings.defaultLanguage === s.$root.lang ? "" : `#?lang=${s.$root.lang}`
-        const modeParam = modeId === "default" ? "" : `?mode=${modeId}`
-        return [location.pathname, modeParam, langParam]
-    }
-
-    s.show_modal = false
-
-    let modal = null
-    utils.setupHash(s, [
-        {
-            key: "display",
-            scope_name: "show_modal",
-            post_change(val) {
-                if (val) {
-                    showModal(val)
                 } else {
-                    if (modal != null) {
-                        modal.close()
-                    }
-                    modal = null
+                    s.openErrorModal({
+                        content: html`<span class="mr-1">{{'login_needed_for_corpora' | loc:lang}}:</span
+                            >${loginNeededHTML()}`,
+                        onClose: () => {
+                            s.waitForLogin = true
+                            statemachine.send("LOGIN_NEEDED", { loginNeededFor })
+                        },
+                    })
                 }
-            },
-        },
-    ])
-
-    const closeModals = function () {
-        s.login_err = false
-        s.show_modal = false
-    }
-
-    var showModal = function (key) {
-        const tmpl = {
-            about: require("../markup/about.html"),
-            login: require("../markup/login.html"),
-        }[key]
-        const params = {
-            template: tmpl,
-            scope: s,
-            windowClass: key,
-        }
-        if (key === "login") {
-            params.size = "sm"
-        }
-        modal = $uibModal.open(params)
-
-        modal.result.then(
-            () => closeModals(),
-            () => closeModals()
-        )
-    }
-
-    s.clickX = () => closeModals()
-
-    s.loggedIn = false
-    const creds = jStorage.get("creds")
-    if (creds) {
-        util.setLogin()
-        s.loggedIn = true
-        s.username = authenticationProxy.loginObj.name
-    }
-    s.loginSubmit = function (usr, pass, saveLogin) {
-        s.login_err = false
-        authenticationProxy
-            .makeRequest(usr, pass, saveLogin)
-            .done(function () {
-                util.setLogin()
-                safeApply(s, function () {
-                    s.show_modal = null
-                    s.restorePreLoginState()
-                    s.loggedIn = true
-                    s.username = usr
+            } else if (!selectedIds.every((r) => allCorpusIds.includes(r))) {
+                s.openErrorModal({
+                    content: `{{'corpus_not_available' | loc:lang}}`,
+                    onClose: () => {
+                        let newIds = selectedIds.filter((corpusId) => allCorpusIds.includes(corpusId))
+                        if (newIds.length == 0) {
+                            newIds = settings["preselected_corpora"]
+                        }
+                        initialzeCorpusSelection(newIds)
+                    },
                 })
+            } else {
+                // here $timeout must be used so that message is not sent before all controllers/componenters are initialized
+                settings.corpusListing.select(selectedIds)
+                $timeout(() => $rootScope.$broadcast("initialcorpuschooserchange", selectedIds), 0)
+            }
+        }
+
+        // TODO the top bar could show even though the modal is open,
+        // thus allowing switching modes or language when an error has occured.
+        s.openErrorModal = ({ content, resolvable = true, onClose = null, buttonText = null, translations = {} }) => {
+            const s = $rootScope.$new(true)
+
+            const useCustomButton = !_.isEmpty(buttonText)
+
+            const modal = $uibModal.open({
+                template: html` <div class="modal-body" ng-class="{'mt-10' : resolvable }">
+                    <div>${content}</div>
+                    <div class="ml-auto mr-0 w-fit">
+                        <button
+                            ng-if="${resolvable}"
+                            ng-click="closeModal()"
+                            class="btn bg-blue-500 text-white font-bold mt-3"
+                        >
+                            <span ng-if="${useCustomButton}">{{'${buttonText}' | loc:lang }}</span>
+                            <span ng-if="!${useCustomButton}">OK</span>
+                        </button>
+                    </div>
+                </div>`,
+                scope: s,
+                size: "md",
+                backdrop: "static",
+                keyboard: false,
             })
-            .fail(function () {
-                c.log("login fail")
-                safeApply(s, () => {
-                    s.login_err = true
-                })
-            })
+
+            s.translations = translations
+
+            s.closeModal = () => {
+                if (onClose && resolvable) {
+                    modal.close()
+                    onClose()
+                }
+            }
+        }
+
+        function getCorporaFromHash() {
+            let selectedIds
+            let { corpus } = $location.search()
+            if (corpus) {
+                selectedIds = corpus.split(",")
+            } else {
+                selectedIds = settings["preselected_corpora"] || []
+            }
+            return selectedIds
+        }
+
+        statemachine.listen("login", function () {
+            if (s.waitForLogin) {
+                s.waitForLogin = false
+                initialzeCorpusSelection(getCorporaFromHash())
+            }
+        })
+
+        initialzeCorpusSelection(getCorporaFromHash())
+    },
+])
+
+korpApp.filter("trust", ["$sce", ($sce) => (input) => $sce.trustAsHtml(input)])
+korpApp.filter("prettyNumber", () => (input) => new Intl.NumberFormat(lang).format(input))
+angular.module("korpApp").filter("maxLength", function () {
+    return function (val) {
+        return val.length > 39 ? val.slice(0, 36) + "…" : val
     }
 })
-
-korpApp.filter("trust", ($sce) => (input) => $sce.trustAsHtml(input))
