@@ -2,13 +2,16 @@
 // SB-newsdesk 1.0b
 // Requirements: JQuery, JQuery.ui.position, trust filter, loc filter, Font Awesome
 
-angular.module("newsdesk", []).directive("newsDesk", ($window, $document, $rootElement, $http, $location) => ({
-    template: `
+angular.module("newsdesk", []).directive("newsDesk", [
+    "$rootElement",
+    "$location",
+    ($rootElement, $location) => ({
+        template: `
             <div>
                 <div ng-if="shouldUseThis" class="newsdesk-opener" 
                      ng-click="togglePopover($event)" 
                      ng-class="{'newsdesk-new-news': numNewNews != 0, 'newsdesk-no-new-news' : numNewNews == 0}">
-                    <i class="fa fa-bell newsdesk-bell"></i>
+                    <i class="fa-solid fa-bell newsdesk-bell"></i>
                     <div class="newsdesk-arrow-box">
                         <span>{{numNewNews}}</span>
                     </div>&nbsp;
@@ -26,102 +29,101 @@ angular.module("newsdesk", []).directive("newsDesk", ($window, $document, $rootE
                     </div>
                 </div>
         </div>`,
-    restrict: "EA",
-    replace: true,
-    scope: { header: "=", storage: "=" },
-    link(scope, elem, attr) {
-        const s = scope
-        s.shouldUseThis = settings.newsDeskUrl != null
+        restrict: "EA",
+        replace: true,
+        scope: { header: "=", storage: "=" },
+        link(scope, elem, attr) {
+            const s = scope
+            s.shouldUseThis = settings["news_desk_url"] != null
 
-        if (!s.shouldUseThis) {
-            return
-        }
-
-        s.onPopoverClick = (event) => event.stopPropagation()
-
-        s.newsitems = []
-        function initData() {
-            let d
-            s.lastChecked = localStorage.getItem(s.storage)
-            if (!s.lastChecked) {
-                d = new Date()
-                d.setFullYear(d.getFullYear() - 1)
-                s.lastChecked = d.toISOString().slice(0, 10)
+            if (!s.shouldUseThis) {
+                return
             }
-            $.ajax({
-                type: "GET",
-                url: settings.newsDeskUrl,
-                async: false,
-                jsonpCallback: "newsdata",
-                contentType: "application/json",
-                dataType: "jsonp",
-                success(json) {
-                    const currentDate = new Date().toISOString().slice(0, 10)
-                    s.newsitems = _.filter(json, (newsitem) => {
-                        return !newsitem.e || newsitem.e >= currentDate
-                    })
-                    let n = 0
-                    for (let nItem of s.newsitems) {
-                        if (nItem.d > s.lastChecked) {
-                            n += 1
+
+            s.onPopoverClick = (event) => event.stopPropagation()
+
+            s.newsitems = []
+            function initData() {
+                let d
+                s.lastChecked = localStorage.getItem(s.storage)
+                if (!s.lastChecked) {
+                    d = new Date()
+                    d.setFullYear(d.getFullYear() - 1)
+                    s.lastChecked = d.toISOString().slice(0, 10)
+                }
+                $.ajax({
+                    type: "GET",
+                    url: settings["news_desk_url"],
+                    async: false,
+                    jsonpCallback: "newsdata",
+                    contentType: "application/json",
+                    dataType: "jsonp",
+                    success(json) {
+                        const currentDate = new Date().toISOString().slice(0, 10)
+                        s.newsitems = _.filter(json, (newsitem) => {
+                            return !newsitem.e || newsitem.e >= currentDate
+                        })
+                        let n = 0
+                        for (let nItem of s.newsitems) {
+                            if (nItem.d > s.lastChecked) {
+                                n += 1
+                            }
                         }
-                    }
 
-                    safeApply(s, () => (s.numNewNews = n))
-                },
+                        safeApply(s, () => (s.numNewNews = n))
+                    },
 
-                error(e) {
-                    console.log("error, couldn't fetch news", e.message)
-                },
-            })
-        }
-
-        s.currentLang = $location.search().lang || "sv"
-
-        s.numNewNews = 0
-        initData()
-
-        s.togglePopover = function (event) {
-            if (s.isPopoverVisible) {
-                s.popHide()
-            } else {
-                s.currentLang = $location.search().lang || "sv"
-                s.popShow()
-                s.numNewNews = 0
+                    error(e) {
+                        console.log("error, couldn't fetch news", e.message)
+                    },
+                })
             }
-            event.preventDefault()
-            event.stopPropagation()
-        }
 
-        const popover = $(".newsdesk-popover")
-        s.isPopoverVisible = false
+            s.numNewNews = 0
+            initData()
 
-        const handleEscape = function (event) {
-            if (event.which === 27) {
-                s.popHide()
-                return false
+            s.togglePopover = function (event) {
+                if (s.isPopoverVisible) {
+                    s.popHide()
+                } else {
+                    s.currentLang = { eng: "en", swe: "sv" }[$location.search().lang || settings["default_language"]]
+                    s.popShow()
+                    s.numNewNews = 0
+                }
+                event.preventDefault()
+                event.stopPropagation()
             }
-        }
 
-        s.popShow = function () {
-            s.isPopoverVisible = true
-
-            popover.show().focus().position({
-                my: "right top",
-                at: "right-10 top+10",
-                of: window,
-            })
-            $rootElement.on("keydown", handleEscape)
-            $rootElement.on("click", s.popHide)
-
-            localStorage.setItem(s.storage, s.newsitems[0].d)
-        }
-
-        s.popHide = function () {
+            const popover = $(".newsdesk-popover")
             s.isPopoverVisible = false
-            popover.hide()
-            $rootElement.off("keydown", handleEscape)
-            $rootElement.off("click", s.popHide)
-        }
-    },
-}))
+
+            const handleEscape = function (event) {
+                if (event.which === 27) {
+                    s.popHide()
+                    return false
+                }
+            }
+
+            s.popShow = function () {
+                s.isPopoverVisible = true
+
+                popover.show().focus().position({
+                    my: "right top",
+                    at: "right-10 top+10",
+                    of: window,
+                })
+                $rootElement.on("keydown", handleEscape)
+                $rootElement.on("click", s.popHide)
+
+                localStorage.setItem(s.storage, s.newsitems[0].d)
+            }
+
+            s.popHide = function () {
+                s.isPopoverVisible = false
+                popover.hide()
+                $rootElement.off("keydown", handleEscape)
+                $rootElement.off("click", s.popHide)
+            }
+        },
+    }),
+])
